make mail templates typesafe
